Rename colors.rs to styling.rs
