+++ src/ui/help/help_widget.rs
@@ -4,8 +4,8 @@ use crate::components::help::{HEADER, RowType, build_docs, format_row};
 use crate::config::TomlFileStore;
 use crate::state::app_settings::AppSettings;
 use crate::state::settings_editor::{SettingsEditorState, SettingsFocus};
-use crate::ui::color::{dim_style, header_style, selected_style};
 use crate::ui::help::settings_panel::{render_picker, render_settings};
+use crate::ui::styling::{dim_style, header_style, selected_style};
 use tui::layout::{Constraint, Flex, Layout};
 use tui::prelude::*;
 use tui::widgets::{Paragraph, Row, Table, TableState};
